+++ packages/backend/src/services/DashboardService/DashboardService.ts
@@ -23,9 +23,9 @@ import {
 import * as Sentry from '@sentry/node';
 import cronstrue from 'cronstrue';
 import { v4 as uuidv4 } from 'uuid';
-import { analytics } from '../../analytics/client';
 import {
     CreateDashboardOrVersionEvent,
+    LightdashAnalytics,
     SchedulerDashboardUpsertEvent,
 } from '../../analytics/LightdashAnalytics';
 import { schedulerClient, slackClient } from '../../clients/clients';
@@ -39,7 +39,8 @@ import { SpaceModel } from '../../models/SpaceModel';
 import { SavedChartService } from '../SavedChartsService/SavedChartService';
 import { hasSpaceAccess } from '../SpaceService/SpaceService';
 
-type Dependencies = {
+type DashboardServiceArguments = {
+    analytics: LightdashAnalytics;
     dashboardModel: DashboardModel;
     spaceModel: SpaceModel;
     analyticsModel: AnalyticsModel;
@@ -49,6 +50,8 @@ type Dependencies = {
 };
 
 export class DashboardService {
+    analytics: LightdashAnalytics;
+
     dashboardModel: DashboardModel;
 
     spaceModel: SpaceModel;
@@ -62,13 +65,15 @@ export class DashboardService {
     savedChartModel: SavedChartModel;
 
     constructor({
+        analytics,
         dashboardModel,
         spaceModel,
         analyticsModel,
         pinnedListModel,
         schedulerModel,
         savedChartModel,
-    }: Dependencies) {
+    }: DashboardServiceArguments) {
+        this.analytics = analytics;
         this.dashboardModel = dashboardModel;
         this.spaceModel = spaceModel;
         this.analyticsModel = analyticsModel;
@@ -112,7 +117,7 @@ export class DashboardService {
                 const deletedChart = await this.savedChartModel.delete(
                     chart.uuid,
                 );
-                analytics.track({
+                this.analytics.track({
                     event: 'saved_chart.deleted',
                     userId: user.userUuid,
                     properties: {
@@ -194,7 +199,7 @@ export class DashboardService {
             dashboard.uuid,
             user.userUuid,
         );
-        analytics.track({
+        this.analytics.track({
             event: 'dashboard.view',
             userId: user.userUuid,
             properties: {
@@ -263,7 +268,7 @@ export class DashboardService {
             user,
             projectUuid,
         );
-        analytics.track({
+        this.analytics.track({
             event: 'dashboard.created',
             userId: user.userUuid,
             properties: DashboardService.getCreateEventProperties(newDashboard),
@@ -327,7 +332,7 @@ export class DashboardService {
                                     },
                                 },
                             );
-                        analytics.track({
+                        this.analytics.track({
                             event: 'saved_chart.created',
                             userId: user.userUuid,
                             properties: {
@@ -365,13 +370,13 @@ export class DashboardService {
 
         const dashboardProperties =
             DashboardService.getCreateEventProperties(newDashboard);
-        analytics.track({
+        this.analytics.track({
             event: 'dashboard.created',
             userId: user.userUuid,
             properties: { ...dashboardProperties, duplicated: true },
         });
 
-        analytics.track({
+        this.analytics.track({
             event: 'duplicated_dashboard_created',
             userId: user.userUuid,
             properties: {
@@ -422,7 +427,7 @@ export class DashboardService {
                 },
             );
 
-            analytics.track({
+            this.analytics.track({
                 event: 'dashboard.updated',
                 userId: user.userUuid,
                 properties: {
@@ -454,7 +459,7 @@ export class DashboardService {
                 user,
                 existingDashboard.projectUuid,
             );
-            analytics.track({
+            this.analytics.track({
                 event: 'dashboard_version.created',
                 userId: user.userUuid,
                 properties:
@@ -504,7 +509,7 @@ export class DashboardService {
             existingDashboard.projectUuid,
         );
 
-        analytics.track({
+        this.analytics.track({
             event: 'pinned_list.updated',
             userId: user.userUuid,
             properties: {
@@ -547,7 +552,7 @@ export class DashboardService {
             );
         }
 
-        analytics.track({
+        this.analytics.track({
             event: 'dashboard.updated_multiple',
             userId: user.userUuid,
             properties: {
@@ -578,7 +583,7 @@ export class DashboardService {
         const deletedDashboard = await this.dashboardModel.delete(
             dashboardUuid,
         );
-        analytics.track({
+        this.analytics.track({
             event: 'dashboard.deleted',
             userId: user.userUuid,
             properties: {
@@ -641,7 +646,7 @@ export class DashboardService {
                         : 0,
             },
         };
-        analytics.track(createSchedulerData);
+        this.analytics.track(createSchedulerData);
 
         await slackClient.joinChannels(
             user.organizationUuid,