+++ packages/backend/src/services/SchedulerService/SchedulerService.ts
@@ -18,8 +18,8 @@ import {
     UpdateSchedulerAndTargetsWithoutId,
 } from '@lightdash/common';
 import cronstrue from 'cronstrue';
-import { analytics } from '../../analytics/client';
 import {
+    LightdashAnalytics,
     SchedulerDashboardUpsertEvent,
     SchedulerUpsertEvent,
 } from '../../analytics/LightdashAnalytics';
@@ -35,8 +35,9 @@ import { SchedulerModel } from '../../models/SchedulerModel';
 import { SpaceModel } from '../../models/SpaceModel';
 import { hasSpaceAccess } from '../SpaceService/SpaceService';
 
-type ServiceDependencies = {
+type SchedulerServiceArguments = {
     lightdashConfig: LightdashConfig;
+    analytics: LightdashAnalytics;
     schedulerModel: SchedulerModel;
 
     dashboardModel: DashboardModel;
@@ -49,6 +50,8 @@ type ServiceDependencies = {
 export class SchedulerService {
     lightdashConfig: LightdashConfig;
 
+    analytics: LightdashAnalytics;
+
     schedulerModel: SchedulerModel;
 
     dashboardModel: DashboardModel;
@@ -59,12 +62,14 @@ export class SchedulerService {
 
     constructor({
         lightdashConfig,
+        analytics,
         schedulerModel,
         dashboardModel,
         savedChartModel,
         spaceModel,
-    }: ServiceDependencies) {
+    }: SchedulerServiceArguments) {
         this.lightdashConfig = lightdashConfig;
+        this.analytics = analytics;
         this.schedulerModel = schedulerModel;
         this.dashboardModel = dashboardModel;
         this.savedChartModel = savedChartModel;
@@ -213,7 +218,7 @@ export class SchedulerService {
                 }),
             },
         };
-        analytics.track(updateSchedulerEventData);
+        this.analytics.track(updateSchedulerEventData);
         await slackClient.joinChannels(
             user.organizationUuid,
             SchedulerModel.getSlackChannels(scheduler.targets),
@@ -265,7 +270,7 @@ export class SchedulerService {
         await this.schedulerModel.deleteScheduler(schedulerUuid);
         await this.schedulerModel.deleteScheduledLogs(schedulerUuid);
 
-        analytics.track({
+        this.analytics.track({
             userId: user.userUuid,
             event: 'scheduler.deleted',
             properties: {
@@ -327,7 +332,7 @@ export class SchedulerService {
             projectUuid,
         );
 
-        analytics.track({
+        this.analytics.track({
             userId: user.userUuid,
             event: 'scheduled_deliveries.dashboard_viewed',
             properties: {