+++ packages/backend/src/services/ShareService/ShareService.ts
@@ -6,11 +6,12 @@ import {
     ShareUrl,
 } from '@lightdash/common';
 import { nanoid as nanoidGenerator } from 'nanoid';
-import { analytics } from '../../analytics/client';
+import { LightdashAnalytics } from '../../analytics/LightdashAnalytics';
 import { LightdashConfig } from '../../config/parseConfig';
 import { ShareModel } from '../../models/ShareModel';
 
-type Dependencies = {
+type ShareServiceArguments = {
+    analytics: LightdashAnalytics;
     shareModel: ShareModel;
     lightdashConfig: Pick<LightdashConfig, 'siteUrl'>;
 };
@@ -18,11 +19,14 @@ type Dependencies = {
 export class ShareService {
     private readonly lightdashConfig: Pick<LightdashConfig, 'siteUrl'>;
 
+    private readonly analytics: LightdashAnalytics;
+
     private readonly shareModel: ShareModel;
 
-    constructor(dependencies: Dependencies) {
-        this.lightdashConfig = dependencies.lightdashConfig;
-        this.shareModel = dependencies.shareModel;
+    constructor(args: ShareServiceArguments) {
+        this.lightdashConfig = args.lightdashConfig;
+        this.analytics = args.analytics;
+        this.shareModel = args.shareModel;
     }
 
     private shareUrlWithHost(shareUrl: ShareUrl) {
@@ -51,7 +55,7 @@ export class ShareService {
         ) {
             throw new ForbiddenError();
         }
-        analytics.track({
+        this.analytics.track({
             userId: user.userUuid,
             event: 'share_url.used',
             properties: {
@@ -78,7 +82,7 @@ export class ShareService {
             createdByUserUuid: user.userUuid,
         });
 
-        analytics.track({
+        this.analytics.track({
             userId: user.userUuid,
             event: 'share_url.created',
             properties: {