+++ packages/backend/src/services/UserService.ts
@@ -35,7 +35,7 @@ import {
 import { randomInt } from 'crypto';
 import { nanoid } from 'nanoid';
 import refresh from 'passport-oauth2-refresh';
-import { analytics, identifyUser } from '../analytics/client';
+import { LightdashAnalytics } from '../analytics/LightdashAnalytics';
 import EmailClient from '../clients/EmailClient/EmailClient';
 import { LightdashConfig } from '../config/parseConfig';
 import Logger from '../logging/logger';
@@ -51,9 +51,11 @@ import { PasswordResetLinkModel } from '../models/PasswordResetLinkModel';
 import { SessionModel } from '../models/SessionModel';
 import { UserModel } from '../models/UserModel';
 import { UserWarehouseCredentialsModel } from '../models/UserWarehouseCredentials/UserWarehouseCredentialsModel';
+import { postHogClient } from '../postHog';
 
-type UserServiceDependencies = {
+type UserServiceArguments = {
     lightdashConfig: LightdashConfig;
+    analytics: LightdashAnalytics;
     inviteLinkModel: InviteLinkModel;
     userModel: UserModel;
     groupsModel: GroupsModel;
@@ -71,6 +73,8 @@ type UserServiceDependencies = {
 
 export class UserService {
     private readonly lightdashConfig: LightdashConfig;
+
+    private readonly analytics: LightdashAnalytics;
 
     private readonly inviteLinkModel: InviteLinkModel;
 
@@ -104,6 +108,7 @@ export class UserService {
 
     constructor({
         lightdashConfig,
+        analytics,
         inviteLinkModel,
         userModel,
         groupsModel,
@@ -117,8 +122,9 @@ export class UserService {
         personalAccessTokenModel,
         organizationAllowedEmailDomainsModel,
         userWarehouseCredentialsModel,
-    }: UserServiceDependencies) {
+    }: UserServiceArguments) {
         this.lightdashConfig = lightdashConfig;
+        this.analytics = analytics;
         this.inviteLinkModel = inviteLinkModel;
         this.userModel = userModel;
         this.groupsModel = groupsModel;
@@ -135,6 +141,60 @@ export class UserService {
         this.userWarehouseCredentialsModel = userWarehouseCredentialsModel;
     }
 
+    private identifyUser(
+        user: LightdashUser & { isMarketingOptedIn?: boolean },
+    ): void {
+        if (this.lightdashConfig.mode === LightdashMode.DEMO) {
+            return;
+        }
+        this.analytics.identify({
+            userId: user.userUuid,
+            traits: user.isTrackingAnonymized
+                ? { is_tracking_anonymized: user.isTrackingAnonymized }
+                : {
+                      email: user.email,
+                      first_name: user.firstName,
+                      last_name: user.lastName,
+                      is_tracking_anonymized: user.isTrackingAnonymized,
+                      is_marketing_opted_in: user.isMarketingOptedIn,
+                  },
+        });
+
+        postHogClient?.identify({
+            distinctId: user.userUuid,
+            properties: {
+                uuid: user.userUuid,
+                ...(user.isTrackingAnonymized
+                    ? {}
+                    : {
+                          email: user.email,
+                          first_name: user.firstName,
+                          last_name: user.lastName,
+                      }),
+            },
+        });
+
+        if (user.organizationUuid) {
+            this.analytics.group({
+                userId: user.userUuid,
+                groupId: user.organizationUuid,
+                traits: {
+                    name: user.organizationName,
+                },
+            });
+
+            postHogClient?.groupIdentify({
+                groupType: 'organization',
+                groupKey: user.organizationUuid,
+                properties: {
+                    uuid: user.organizationUuid,
+                    name: user.organizationName,
+                },
+                distinctId: user.userUuid,
+            });
+        }
+    }
+
     private async tryVerifyUserEmail(
         user: LightdashUser,
         email: string,
@@ -144,7 +204,7 @@ export class UserService {
             email,
         );
         if (updatedEmails.length > 0) {
-            analytics.track({
+            this.analytics.track({
                 userId: user.userUuid,
                 event: 'user.verified',
                 properties: {
@@ -185,8 +245,8 @@ export class UserService {
             activateUser,
         );
         await this.inviteLinkModel.deleteByCode(inviteLink.inviteCode);
-        identifyUser(user);
-        analytics.track({
+        this.identifyUser(user);
+        this.analytics.track({
             event: 'user.created',
             userId: user.userUuid,
             properties: {
@@ -223,7 +283,7 @@ export class UserService {
         await this.sessionModel.deleteAllByUserUuid(userUuidToDelete);
 
         await this.userModel.delete(userUuidToDelete);
-        analytics.track({
+        this.analytics.track({
             event: 'user.deleted',
             userId: user.userUuid,
             properties: {
@@ -289,13 +349,13 @@ export class UserService {
             userUuid,
         );
         await this.emailClient.sendInviteEmail(user, inviteLink);
-        analytics.track({
+        this.analytics.track({
             userId: user.userUuid,
             event: 'invite_link.created',
         });
 
         const organization = await this.organizationModel.get(organizationUuid);
-        analytics.track({
+        this.analytics.track({
             userId: user.userUuid,
             event: 'permission.updated',
             properties: {
@@ -323,7 +383,7 @@ export class UserService {
             throw new NotExistsError('Organization not found');
         }
         await this.inviteLinkModel.deleteByOrganization(organizationUuid);
-        analytics.track({
+        this.analytics.track({
             userId: user.userUuid,
             event: 'invite_link.all_revoked',
         });
@@ -340,7 +400,7 @@ export class UserService {
                 updatedGroups.map(async (groupUuid) => {
                     const updatedGroup =
                         await this.groupsModel.getGroupWithMembers(groupUuid);
-                    analytics.track({
+                    this.analytics.track({
                         event: 'group.updated',
                         userId: data[0].userUuid,
                         properties: {
@@ -392,8 +452,8 @@ export class UserService {
                 refreshToken,
             });
             await this.tryVerifyUserEmail(loginUser, openIdUser.openId.email);
-            identifyUser(loginUser);
-            analytics.track({
+            this.identifyUser(loginUser);
+            this.analytics.track({
                 userId: loginUser.userUuid,
                 event: 'user.logged_in',
                 properties: {
@@ -428,7 +488,7 @@ export class UserService {
                 refreshToken,
             });
             await this.tryVerifyUserEmail(sessionUser, openIdUser.openId.email);
-            analytics.track({
+            this.analytics.track({
                 userId: sessionUser.userUuid,
                 event: 'user.identity_linked',
                 properties: {
@@ -517,7 +577,7 @@ export class UserService {
             await this.organizationModel.update(user.organizationUuid, {
                 name: organizationName,
             });
-            analytics.track({
+            this.analytics.track({
                 userId: user.userUuid,
                 event: 'organization.updated',
                 properties: {
@@ -559,8 +619,8 @@ export class UserService {
             },
         );
 
-        identifyUser(completeUser);
-        analytics.track({
+        this.identifyUser(completeUser);
+        this.analytics.track({
             event: 'user.updated',
             userId: completeUser.userUuid,
             properties: {
@@ -588,7 +648,7 @@ export class UserService {
             openIdentity.issuer,
             openIdentity.email,
         );
-        analytics.track({
+        this.analytics.track({
             userId: user.userUuid,
             event: 'user.identity_removed',
             properties: {
@@ -626,8 +686,8 @@ export class UserService {
                 email,
                 password,
             );
-            identifyUser(user);
-            analytics.track({
+            this.identifyUser(user);
+            this.analytics.track({
                 userId: user.userUuid,
                 event: 'user.logged_in',
                 properties: {
@@ -664,7 +724,7 @@ export class UserService {
         } else {
             await this.userModel.createPassword(user.userId, data.newPassword);
         }
-        analytics.track({
+        this.analytics.track({
             userId: user.userUuid,
             event: 'password.updated',
         });
@@ -679,8 +739,8 @@ export class UserService {
             user.email,
             data,
         );
-        identifyUser(updatedUser);
-        analytics.track({
+        this.identifyUser(updatedUser);
+        this.analytics.track({
             userId: updatedUser.userUuid,
             event: 'user.updated',
             properties: updatedUser,
@@ -718,11 +778,11 @@ export class UserService {
             throw new ForbiddenError('Password credentials are not allowed');
         }
         const user = await this.userModel.createUser(createUser);
-        identifyUser({
+        this.identifyUser({
             ...user,
             isMarketingOptedIn: user.isMarketingOptedIn,
         });
-        analytics.track({
+        this.analytics.track({
             event: 'user.created',
             userId: user.userUuid,
             properties: {
@@ -732,7 +792,7 @@ export class UserService {
             },
         });
         if (isOpenIdUser(createUser)) {
-            analytics.track({
+            this.analytics.track({
                 userId: user.userUuid,
                 event: 'user.identity_linked',
                 properties: {
@@ -767,7 +827,7 @@ export class UserService {
                 expiresAt,
                 data.email,
             );
-            analytics.track({
+            this.analytics.track({
                 userId: user.userUuid,
                 event: 'password_reset_link.created',
             });
@@ -787,7 +847,7 @@ export class UserService {
                 data.newPassword,
             );
             await this.passwordResetLinkModel.deleteByCode(link.code);
-            analytics.track({
+            this.analytics.track({
                 userId: user.userUuid,
                 event: 'password_reset_link.used',
             });
@@ -968,7 +1028,7 @@ export class UserService {
                 : undefined,
         );
 
-        await analytics.track({
+        await this.analytics.track({
             userId: user.userUuid,
             event: 'user.joined_organization',
             properties: {
@@ -1038,7 +1098,7 @@ export class UserService {
                 user.userUuid,
                 data,
             );
-        analytics.track({
+        this.analytics.track({
             userId: user.userUuid,
             event: 'user_warehouse_credentials.created',
             properties: {
@@ -1061,7 +1121,7 @@ export class UserService {
             userWarehouseCredentialsUuid,
             data,
         );
-        analytics.track({
+        this.analytics.track({
             userId: user.userUuid,
             event: 'user_warehouse_credentials.updated',
             properties: {
@@ -1082,7 +1142,7 @@ export class UserService {
             user.userUuid,
             userWarehouseCredentialsUuid,
         );
-        analytics.track({
+        this.analytics.track({
             userId: user.userUuid,
             event: 'user_warehouse_credentials.deleted',
             properties: {